pyproject.toml
[tool.poetry]
name = "rc-circuit"
version = "0.1.0"
description = ""
authors = ["Nicolás Pozas <Pozas91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
rpi-gpio = "^0.7.1"
pi-rc522 = "^2.3.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
pymongo = "^4.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
